Fix encoding to Ogg Theora with Vorbis.
This applies to Kdenlive bug 465. After fixing the huge memory leak,
there was an audio sync problem, and the fix for that might help other
formats as well (fix was to initialize the audio codec time base). This
also increases the size of the audio encoding buffer to fix flac
encoding.

Signed-off-by: Dan Dennedy <dan@dennedy.org>