Patch from Ivan Schreter to fix seeking on AVCHD.
This is intended to improve seeking in general, but it not ready for all
formats. It can be explicitly enabled by setting the new_seek property
to 1. I added code to enable it by default for H.264 in MPEG-2 Transport
Stream, but one can disable it by setting new_seek to 0.

Signed-off-by: Dan Dennedy <dan@dennedy.org>