Invert position of normalization filters with framebuffer producer.
Previously, the framebuffer producer loaded a normalized producer, which
can give undesired results with things like crop and really any time you
want to filter the unpadded images of the speed-altered video. Now, the
framebuffer uses the new "abnormal" producer to load the clip without
normalization filters and sets appropriate frame properties to allow the
normalizing filters attached to the framebuffer to act appropriately.
This new abnormal filter is simply an alias to the existing loader
filter, which uses the name by which it is invoked to toggle the
behaviour of whether to attach normalizing filters.