Fix seek on some AVCHD causing infinite loop.
This detects when libavformat starts to consistently report invalid PTS
and then will disable Ivan's new_seek code. In the example file I have
this only happens when seeking backwards, which typically implies an
interactive use case (except when using framebuffer with speed < 0).
Then, it should still use the new_seek when simply apply an in point to
allow for a clean, accurate cut.