Add producer variant avformat-novalidate.
The purpose of this is to increase the speed of loading playlists with
known good files. Use with care. This assumes a few properties have been
set, in particular "length." This was only tested thus far by modifying
the output of consumer xml to change mlt_service from "avformat" to
"avformat-novalidate".