Add fetching previous and next frames in producers.
This is only enabled when the property need-previous-next is set true on
the producer. This also adds firing a service-changed event on the
filter when it gets attached so the filter can set this property on the
producer to which it is attached. These frame references are set as
"previous frame" and "next frame" properties on the current frame. It is
also important to note that these frames do not have ANY filters applied
to them, which is important for YADIF and telecide filters, which
process before all other filters.