Default SDL to use the onefield deinterlace filter.
The previous default in the deinterlace filter was linearblend. The new
default is yadif. However, onefield is faster is gives cleaner results
than linearblend where preserving most resolution is not a factor. Since
most usess of SDL are applications with preview windows smaller than
actual resolution, it makes sense to use onefield by default in the SDL
consumers.