Multitrack audio encoding continued.
This version changes the configuration and remapping. The number of
channels per output audio track is set using "channels.<N>" properties
on the avformat consumer, where <N> is a 0-based numeric representing
the output track. At this time, all tracks must share all other
attributes such as sample rate, codec, and bitrate.

As for the remapping, this attempts to reuse the meta.map.audio...
properties set on the producers as used with the sdi consumer. One
exception: to skip or silence channels at the beginning tracks or in the
middle, you must add additional map properties to the end of the list to
simulate these "gaps."