Remove audio processing from the worker threads.
This has a bad interaction with the avformat producer, which contains a
buffer of unused decoded samples. This shifts audio processing to the
main consumer thread, which is often light anyways. I recommend to set
the threads property to 2 or more on the avformat consumer to offload
video encoding to separate threads from the audio processing and
encoding.