Cleanup existing native color space conversions.
This change clarifies that the existing conversions are according to the
ITU 601 standard and scaled to and from full gamut RGB. Also, adjust 2
coefficients according to Charles Poynton's matrices. This does not yet
attempt to make any substantial improvements.

Finally, it replaces the verbose logic and redundancy in the image
conversion routine with a concise function dispatch table.