Enhance luma filter to work with animated filters.
Previously, in a slideshow the luma filter would apply the dissolve or
wipe repeatedly over a slide. For example, with a slide duration of 75
frames and a luma period of 25 (expressed as 24), the wipe occurs 3
times. However, since the slides were static, you did not notice it
until the transition at the beginning of a new slide - when you do want
to see it. However, upon adding an affine filter to animate a smooth
pan/zoom, you do notice the extra repetitions - the slides appear to
blend with one another when they are not transitioning.

This change fixes that with new properties 'cycle' and 'duration'. Cycle
is basically a replacement for 'period' that fixes the semantics to
properly represent a duration. Where you would previously express, for
example, period=24, you now say cycle=25. The 'duration' property
prevents the repeating and expresses that the transition should only
occur within the first N frames of the cycle. See demo/mlt_slideshow2
for an example of using it in conjunction with the affine filter!