Get initial skipping of luma scaling to work.
When the avformat producer property skip_luma_scale is set to 1, then we
do not scale the luma on the first YCbCr to RGB conversion. This is only
done once because swscale always downscales luma when converting RGB to
YCbCr, and we need to keep the conversions symmetrical to prevent luma
contraction (loss of contrast).