Add an automatic profile feature to melt.
Here are the main use cases this feature provides:
- Given a regular (non-mlt-xml) media file, melt reads the media
attributes and generates an equivalent MLT profile. This makes it easier
to transcode without changing or specifying resolution, aspect, and
framerate.
- Given a MLT XML file containing a profile attribute or element, melt
loads the specified profile. A composition typically contains profile-
without you having to remember.
- Given a MLT XML containing a profile but also specifying a -profile
option, melt automatically uses the 'consumer' producer with the
requested profiles. This is similar to the above case, but when
explicitly choosing a profile different than the composition one should
use the consumer producer. This just makes melt smarter and more
automatic.