Fix field order correction on cached image.
When the avformat producer is using image caching, the field order is
top-field-first, and the consumer is paused then the field order
correction was applied to the cached image. As a result, when repeating
the image due to being paused, the active image would scroll down the
frame. This fixes it by copying to a new image instead of reusing the
cached image.