Fix serializing xmldata in kdenlivetitle (kdenlive-1841).
Patch below fixes an issue with the kdenlivetitle producer. Basically,
the problem was that when loading a kdenlivetitle from a file, all the
properties were serialized and passed to the xml consumer.

The problem became more obvious with the "embeded" images in titles,
which then caused images to be embedded inside the kdenlive project
file, causing problems like reported in this issue:

http://kdenlive.org/mantis/view.php?id=1841

With the patch, titles loaded from a file will not copy the xmldata.