Synchronize Producer.set_speed(0) with sdl_preview.
This also helps prevent deadlock while waiting for consumer-sdl-paused
event. Not 100% yet, but 100% requires script (swig) apps to handle the
event asynchronously via an event listener, which is not available yet
for most - only ruby. Furthermore, they would really like to be able to
pass opaque data to the asynchronous handler, which is not yet available
in the framework.

A good example here is pausing playback prior to seeking to a specific
frame. The app should be able to make a consumer-paused event handler to
which it can pass the new position, so it can properly seek after the
pause has officially occurred. Without the ability to pass opaque data,
it must save the new position as an instance variable to use within the
handler - once it has support for event listeners that is.