Add 'factor' property to scale filters.
Under certain conditions it can be desirable to manually change the
resolution. Caution: one can still not use this in a completely generic
way with this change.

For example, in a realtime playout situation, one can attach swscale
with factor=0.25, followed by frei0r.cluster, followed by swscale again
with no properties. The first swscale will downscale the image for the
heavy cluster filter. The last swscale will upscale it to make the rest
of the project components happy.