Fix luma semantics when both reverse and invert.
Previously, when not using a wipe (dissolve), invert would make the
transition have no effect. Now, it works and does the same thing as
reverse. Also, when using a wipe, reverse had no effect when invert was
set, and the desired effect could not be achieved. Now, it works as
expected.