Fix sample rate and channel count from avformat producer.
If unable to resample or the decoder does not give the requested channel
count, then we should return the actual values so downstream components
can do the correct thing if they can. A good example of this is 6
channel AAC when the consumer requests 2 channels. Unlike the AC-3
decoder, the AAC decoder can not downmix. So, we were returning audio
pcm containing 6 channels but with a channel count of 2.