Revert commits e49264, c9cb51, and e7a775.
These commits are related to kdenlive bugs 1962 and 2003.
This is about an a/v sync problem when using full duration of a file.
The changes are reverted because I believe the root cause is returning
fewer samples than requested on the last frame of audio, and that was
fixed in the previous commit.
Also, the adjust_length property is removed because it did not work; the
property was not available when trying to get it from within the
constructor. The correct way to adjust the length is to first get the
length and then set the length and out properties.