Ensure serialized yaml is LC_NUMERIC=C.
This is needed to ensure consistency. The metadata is read as string
data and converted to numbers on demand. This is why loading defaults to
C locale - since they were authored with that in mind. However, some
plugins dynamically generate metadata and may set values with int or
double. Therefore, we need to make it all output consistently, and for
now that means C.