Add re-open capability to decklink producer.
This causes the producer to automatically close itself when it reaches
the "end" as defined by an out or length property. Then, the same
producer can be re-opened if reused elsewhere in a playlist or another
instance of the decklink producer can use the same card. This requires
that the decklink producer be permitted to play out to its duration.