make speed of editing AVCHD tolerable
This only works for FFmpeg v0.9.1 or newer but not yet for libav.
This uses the image scaling interpolation mode as a hint to perform less
(nearest) or more (bilinear or bicubic) diligent seeking. Since most
editors are using the SDL consumer with rescale=nearest, this makes
scrubbing and cuts/transitions faster. Then, upon encoding, since the
avformat consumer uses bilinear by default, it will use the slower, more
accurate seeking to ensure the integrity of cuts with an in
point. This change has some quirks: frame-stepping backwards sometimes
does not update the image, and sometimes a big jump forward shows
artifacts.