riff.cc: Initialize data in constructor
Fixes Coverity CID 709444: Uninitialized scalar field (UNINIT_CTOR)
Non-static class member ""length"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""name"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""offset"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""parent"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""type"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""written"" is not initialized in this constructor nor in any functions that it calls.