kino/filehandler.cc: initialize variables
Fixes Coverity CID 709443: Uninitialized scalar field (UNINIT_CTOR)
Non-static class member ""everyNthFrame"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""framesToSkip"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""maxFileSize"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""timeStamp"" is not initialized in this constructor nor in any functions that it calls.