rtaudio/consumer_rtaudio.cpp: initialize variables
Fixes Coverty CID 710879: Uninitialized pointer field (UNINIT_CTOR)

n-static class member ""queue"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.child" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.close" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.event_listener" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.is_stopped" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.local" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.put" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.queue" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.start" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.stop" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.threads" is not initialized in this constructor nor in any functions that it calls.
Non-static class member field "consumer.worker_threads" is not initialized in this constructor nor in any functions that it calls.