Increase libavutil version required to use PIX_FMT_YUVA444P.
PIX_FMT_YUVA444P was added to ffmpeg in version 51.35.101 - and so the precompiler check was correctly set to that version. However, it has not yet been added to libav. And libav recently increased the libavutil version to 51.38.0. This causes a compilation error against libav master because the precompiler check passes, but PIX_FMT_YUVA444P is not defined.

The current libavutil versions are:
 * ffmpeg master: 51.69.100
 * ffmpeg 0.11:   51.54.100
 * libav  master: 51.38.0

This commit sets the precompiler to check against version 51.54.0. This will allow PIX_FMT_YUVA444P to be used when compiled against both ffmpeg master and ffmpeg 0.11 - while avoiding errors when compiling against libav master. However, if libav keeps incrementing the version without adding PIX_FMT_YUVA444P, this number may have to be increased again in the future.