Use $avformat_suffix in pkg-config invocations
It is required to really build MLT against a non-default FFmpeg (i.e.,
makes --avformat-suffix option really work).
The same change should probably be applied to static FFmpeg section,
as well as to a MINGW32_NT-* case, but I'm not able to test this, so
I didn't touch those parts.