More fixes for force_full_luma (kdenlive-2799).
This change lets the image converter downstream of the avformat producer
perform utilize the range as-needed. Then, when the rescale filter sees
that the force_full_range is set on the frame but has not yet been
applied, forces a conversion to RGB to enforce it. In addition, the
recently added force_full_luma property on the avformat producer is
removed because it is redundant with AVOption color_range=2.