Version modules and data directories, and melt
Allow the "extras" of binary incompatible versions of MLT to be installed
simultaneously.
I don't like the idea of versioning the melt binary. But kdenlive is the main
user of MLT and it expects the same formats support from both the libmltX it
is linked to, and the melt binary it uses to do the actual work.