Take MltInput out of the EffectChain.
Having the MltInput be an Input which forwards down to the real implementation
has been a source of multiple headaches, and now lastly, when finalize()
disappeared, source of a broken build. We still need the unified
set_pixel_pointer() etc., but the class is now simply a holder of the Input*,
not a forwarder as viewed from the EffectChain.