Use the shared ResourcePool also for ad-hoc EffectChains.
We sometimes create an EffectChain just to get a temporary MLT image
into Movit. This, like any other, should use the ResourcePool.

A side-effect of this is that it serves as a temporary workaround
for Kdenlive bug 3253, since it delays the destruction of the
ResourcePool until the end of the program.