Bug #MLT-216: Deletion of unnecessary checks before specific function calls
A couple of functions perform input parameter validation before their
implementations will try further actions with side effects. Some calling
functions perform similar safety checks.

Functions which release a system resource are occasionally documented
in the way that they tolerate the passing of a null pointer for example.
There is not a need because of this fact that a function caller repeats
a corresponding check.

A previously added semantic patch approach suggested the changes that are
shown here.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>