Revert trying to coerce XML numeric locale on Windows.
Windows does not use ISO language and country codes for locale names.
This was an attempt to avoid generating XML with locale names not
understandable on other platforms. However, this was causing more
trouble that it was worth. Often, the numeric property values would
become inconsistent (radix character) because either a properties list
had a different locale, or a property had already converted to string
under the system locale. At some point, we probably need a way to
convert Windows locale names to ISO codes.