Add default constructor, const copy constructor and copy operator=.
The default constructor is useful to define frame objects that will be assigned
at a later time. In Mlt++, some objects have a default constructor that
initializes the underlying MLT type while others create an invalid object. The
Frame copy constructor creates an invalid object so that empty frames can be
predeclared without the expense of initializing an empty frame only to destroy
it all when it is actually assigned to a frame.

The const copy constructor duplicates the non-const constructor, but is useful
so that frames can be initialized from tempory variables.

The copy operator= is necessary so that frames can be stored in containers.

Additionally, with this change, Mlt::Frame implements "the big three" which
is a good practice.