Restore src alpha on "burned" image.
The filter reset all alpha values to 0. This isn't a problem if sent directly
to some consumer who ignore alpha values. But when passed to a transition as a
"b" frame, the content would be completely transparent, thus showing only the
"a" frame.

Test case:
melt -video-track color:red -video-track frei0r.test_pat_G -filter burningtv
-transition composite fill=1

Before this change, the output would be red frame. After this fix, the output
is the pattern with fire on the squares - and the red frame completly obscured.