Improve A/V sync when starting/stopping/refreshing.
When stopping, the consmer would throw away frames that were in the video queue
rather than provide them to the application.

Additionally, in some cases, while stopping, the consumer thread would output a
newer frame and then the video thread would output an older frame - causing
frames to be provided out of order.

Also, when refreshing, frequently two frames would be displayed rather than
one.

The fundamental principle of this change is that if the audio was displayed for
a frame, the video should also be provided for the display - even when
stopping.