Render audio before video.
This provides better support for audio visualization filters because the filter
has no other way to know the audio channels and sample rate that will be
requested by the consumer when it is processing the video frame. With this
change, the audio is always pre-rendered when the visualization's get_image
function is called.