Add -abort option to melt.
This makes melt quit without cleanup at the end of processing.
This is useful in certain difficuly-to-tame combinations of
code, dependencies, build, and OS where the output is okay,
but a failure may occur during cleanup of data structures and
resources. Sometimes, you want to tell melt to simply abort at the
end instead of cleaning up.