Initialize some Movit structs.
I received notice from Steinar that YCbCrFormat is going to use a new
field that will use its default if the numeric field value is zero. We
cannot set the field yet because we do not have version detection and
don't want to break compile on older, release versions of Movit.