Fix loading a second LADSPA plugin.
When using Flowblade or OpenShot to edit videos, trying to apply a
second LADSPA effect from the swh-tools package to an audio stream
caused a coredump in MLT's jackrack module.

After some debugging, this proved to be lack of a null pointer check in
jackrack: the null pointer was unexpected, since it corresponded to the
LADSPA plugin having been incorrectly loaded.

After fixing MLT jackrack to not dump core and fixing up the logging,
the problem with LADSPA loading was traced to unnecessarily using
RTLD_GLOBAL in the dlopen() call for the LADSPA plugin in one of two
locations where the plugin was opened. This (for some reason I don't
understand) prevented the swh_init constructor for the LADSPA plugin
from being called at load time, which in turn caused the plugin to fail
to be configured properly.

With RTLD_GLOBAL removed, MLT worked as expected. Flowblade applied
multiple swh-plugins LADSPA effects to an audio track successfully.

This patch cleans up the debug logging for MLT jackrack, prevents a
coredump in the case that a LADSPA plugin reports no configuration, and
removes the offending RTLD_GLOBAL.