Fix #214 setlocale in mlt_properties is slow.
On Windows, this change makes mlt_properties_set_lcnumeric() no-op and
mlt_property and mlt_properties insenstive to the locale parameters.
IOW, it always uses the thread's locale. Mainly, this means that the
LC_NUMERIC attribute in MLT XML is not honored on Windows, and the
document is instead interpreted in the thread's locale.