Fix SDL2 regression on macOS on previous commit.
Oddly enough, I got lucky with the first attempt at SDL2 on macOS. ONe
might expect that performing all SDL video functions on the same thread
would be safer and more portable overall. However, with that change,
macOS was giving an error:
!!! BUG: The current event queue and the main event queue are not the
same. Events will not be handled correctly. This is probably because
_TSGetMainThread was called for the first time off the main thread.