Add sdl2 module that may coexist with sdl (1.2).
Now, it is possible to build both SDL 1.2 and SDL 2 consumers.
However, using both in the same process is not supported.
Builders can choose to disable one or the other at configure
time as they deem fit. It can be desirable to build both to allow
an application such as Flowblade that only works with SDL 1.2
while another application such as melt uses SDL 2.