Fix applying channel_layout property.
MLT allows specifying channel_layout as a string representing the
channel layout. But avcodec expects channel_layout to be specified
as an integer which represents a channel position bitmap. This change
stops the channel_layout property from being passed to avcodec so that
the value calculated by MLT will be used instead.