Improve surround support in SDL2 on Windows.
The directsound implementation in SDL does not support > 2 channels. But the
xaudio2 implementation does. This change will try all available drivers to try
to sucessfully open the audio device. If none of the drivers work with surround,
then it will revert to 2 channels.

The user can still request a specific driver using the audio_driver parameter.