Improve surround support in RtAudio on Windows.
The directsound implementation in RtAudio does not support > 2 channels. But the
WASAPI implementation does. This change will try all available drivers to try
to sucessfully open the audio device. If none of the drivers work with surround,
then it will revert to 2 channels.

In the case that the module reverts to 2 channels, the requested number of
channels will still be processed, but only the first two channels will be
output.

These implementation changes were carried over from the SDL2 module to match
the behavior of consumer_sdl2_audio.