Fix producer cache resize heuristic too conservative.
This partially fixes Shotcut bug https://github.com/mltframework/shotcut/
issues/562

There could be several tracks each with a transition that consists of a
transition at the same time where the transition is a nested tractor
consisting of two clips. In addition, there could be a clip containing a
MLT XML virtual clip that contains multiple tracks. This is not a
foolproof solution for the latter, but this heuristic change helps in
the meantime.