Fix improper sync on audio files with discard packets.
Fixes https://github.com/mltframework/shotcut/issues/571

On the sample file provided, I noticed AVPacket timestamps were
fluctuating a lot, which caused the avformat producer to discard many
decoded samples. Then, I noticed that some packets were marked with a
flag AV_PKT_FLAG_DISCARD, which is a somewhat new flag:

Flag is used to discard packets which are required to maintain valid
decoder state but are not required for output and should be dropped
after decoding.

But I noticed none of the ff tools use this flag, and I noticed that it
factors into the got_frame_ptr parameter on avcodec_decode_audio4(). A
packet with this flag yields 0 for got_frame_ptr. Thus, if that parameter
is always 0 for this packet I should consider it discarded and not
factor its PTS into the synchronization logic.