Fix #351 remove call to XInitThreads in the qt module.
Calling it may cause problems with applications that need to call
XInitThreads and is not necessary for window-less operation of melt (at
least with Qt 5) even when using the qglsl consumer with Movit. The man
page clearly says that it "must be the first Xlib function a multi-
threaded program calls," and we have no way to control that within the
MLT module.