Restart SDL audio in case of timeout.
A timeout may occur due to a device or device configuration change,
typically initiated by the user. This change tries to detect it and
restart SDL audio. Note that on Windows while testing, any attempt to
close the SDL audio device or quit the audio subsystem results in
deadlock.

https://forum.shotcut.org/t/changing-the-bit-depth-and-sample-rate-in-
windows-renders-the-player-unresponsive-in-v18-06-02-and-later-
v18-03-06-and-earlier-dont-have-this-problem/7185