Fix crash in clear_property() with mlt_animation.
This happens when switching a rect property between animated and not.
When not animated, the destructor gets set to free(mlt_rect). Then, when
animated, the data pointer was set to the animation object.
Subsequently, in clear_property() both the destructor and
mlt_animation_clear() are called resulting in a double free.