Fix setlocale(LC_ALL) on Windows with env var.
Windows setlocale() does not use the LC_ALL environment variable when
set like other POSIX OS. According to
https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/
setlocale-wsetlocale
"If locale points to an empty string, the locale is the implementation-
defined native environment."
and
"Sets the locale to the default, which is the user-default ANSI code
page obtained from the operating system."
No mention of environment variables, and my own testing confirms. This
change merely provides an additional level of control that appplications
can use to affect MLT locale.